editoast: move hash function to SimulationRequest impl
